Organize form files
