+++ src/components/Modal.js
@@ -1,5 +1,4 @@
 import React from "react";
-import SignupForm from './LoginForm'
 
 const Modal = ({ show, children }) => {
   const showHideClassName = show ? "block" : "none";
